updates to ionosphere and pub results
