+++ README.md
@@ -6,6 +6,9 @@ Bayesian neural networks via MCMC: tutorial
 - `03 Linear Model.ipynb`: Implementing an MCMC algorithm to fit a Bayesian linear regression model
 - `04 Bayesian Neural Network.ipynb`: Implementing an MCMC algorithm to fit a Bayesian neural network
 
+## Acknowledgements
+
+Data Analytics for Resources and Environment (DARE), ARC Industrial Transformation Training Centre, University of Sydney, Sydney, Australia
 
 ## Data citations
 - Abalone: Dua, D. and Graff, C. (2019). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.
